3commit

+++ src/main/java/ma/thinline/gestion_workflow/controller/RoleController.java
@@ -1,14 +1,13 @@
 package ma.thinline.gestion_workflow.controller;
 
 import ma.thinline.gestion_workflow.dto.RoleDto;
-import ma.thinline.gestion_workflow.dto.UtilisateurDto;
-import ma.thinline.gestion_workflow.modele.Role;
 import ma.thinline.gestion_workflow.service.RoleServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -19,37 +18,46 @@ public class RoleController {
     private RoleServiceImpl roleService;
 
     @GetMapping(value = "listroles")
-    public ResponseEntity<List> ViewRoles() {
-        List<RoleDto> list= roleService.getAllRoles();
-        return  ResponseEntity.ok(list);
+    public ResponseEntity<Object> ViewRoles() {
+        return new ResponseEntity<>(roleService.getAllRoles(), HttpStatus.OK);
     }
 
     @GetMapping(value ="name/{name}")
-    public ResponseEntity<RoleDto> getRoleByName(@PathVariable String name){
-        return ResponseEntity.ok(roleService.getRoleByName(name));
+    public ResponseEntity<Object> getRoleByName(@PathVariable String name){
+        RoleDto dto=roleService.getRoleByName(name);
+        if(dto==null)
+            return new ResponseEntity<>("le role n'existe pas",HttpStatus.OK);
+        return new ResponseEntity<>(dto,HttpStatus.OK);
     }
 
     @GetMapping(value = "id/{id}")
-    public ResponseEntity<RoleDto> getRoleById(@PathVariable Long id){
-        return ResponseEntity.ok(roleService.getUserById(id));
+    public ResponseEntity<Object> getRoleById(@PathVariable Long id){
+        RoleDto dto=roleService.getRoleById(id);
+        if(dto==null)
+            return new ResponseEntity<>("le role n'existe pas",HttpStatus.OK);
+        return new ResponseEntity<>(dto,HttpStatus.OK);
     }
 
     @PostMapping(value = "createrole")
-    public ResponseEntity<String> CreateRole(@RequestBody RoleDto dto){
+    public ResponseEntity<Object> CreateRole(@Valid @RequestBody RoleDto dto){
         roleService.saveRole(dto);
-        return ResponseEntity.ok("le role a été créé avec succés");
+        return new ResponseEntity<>("le role a été créé avec succés",HttpStatus.CREATED);
     }
 
-    @PutMapping(value = "editrole/{id}")
-    public  ResponseEntity<String> UpdateRole(@PathVariable Long id, @RequestBody RoleDto dto){
+    @PutMapping(value = "edit/{id}")
+    public  ResponseEntity<Object> UpdateRole(@PathVariable Long id,@Valid @RequestBody RoleDto dto){
+        if(roleService.getRoleById(id)==null)
+            return new ResponseEntity<>("le role n'existe pas",HttpStatus.OK);
         roleService.UpdateRole(id,dto);
-        return ResponseEntity.ok("le role a été modifié avec succés");
+        return new ResponseEntity<>("le role a été modifié avec succés",HttpStatus.OK);
     }
 
     @DeleteMapping(value ="deleterole/{id}")
-    public ResponseEntity<String> DeleteRole(@PathVariable Long id ){
+    public ResponseEntity<Object> DeleteRole(@PathVariable Long id ){
+        if (roleService.getRoleById(id)== null)
+            return new ResponseEntity<>("le role n'existe pas", HttpStatus.OK);
         roleService.DeleteRole(id);
-        return ResponseEntity.ok("le role a été supprimé avec succés");
+        return new ResponseEntity<>("le role a été supprimé avec succés", HttpStatus.OK);
     }
 
 }
